dbyteAPI/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
chardet==3.0.4
click==8.1.7
fastapi==0.115.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==24.1
pdf2image==1.17.0
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pytesseract==0.3.13
python-multipart==0.0.9
rfc3986==1.5.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
